Refactors common unit test code and split frames tests separately
