+++ multibody/tree/body.h
@@ -214,10 +214,14 @@ class Body : public MultibodyTreeElement<Body<T>, BodyIndex> {
   virtual SpatialInertia<T> CalcSpatialInertiaInBodyFrame(
       const MultibodyTreeContext<T>& context) const = 0;
 
+  /// Returns the pose `X_WB` of this body B in the world frame W as a function
+  /// of the state of the model stored in `context`.
   const Isometry3<T>& EvalPoseInWorld(systems::Context<T>& context) const {
     return this->get_parent_tree().EvalBodyPoseInWorld(context, *this);
   }
 
+  /// Returns the spatial velocity `V_WB` of this body B in the world frame W
+  /// as a function of the state of the model stored in `context`.
   const SpatialVelocity<T>& EvalSpatialVelocityInWorld(
       systems::Context<T>& context) const {
     return this->get_parent_tree().EvalBodySpatialVelocityInWorld(
